@@ -24,52 +24,46 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "17.3.12",
-    "@angular/cdk": "17.3.10",
-    "@angular/common": "17.3.12",
-    "@angular/compiler": "17.3.12",
-    "@angular/core": "17.3.12",
-    "@angular/forms": "17.3.12",
-    "@angular/localize": "17.3.12",
-    "@angular/platform-browser": "17.3.12",
-    "@angular/platform-browser-dynamic": "17.3.12",
-    "@angular/router": "17.3.12",
-    "@angular/service-worker": "17.3.12",
-    "@ngx-translate/core": "^14.0.0",
+    "@angular/animations": "18.2.2",
+    "@angular/cdk": "18.2.2",
+    "@angular/common": "18.2.2",
+    "@angular/compiler": "18.2.2",
+    "@angular/core": "18.2.2",
+    "@angular/forms": "18.2.2",
+    "@angular/localize": "18.2.2",
+    "@angular/platform-browser": "18.2.2",
+    "@angular/platform-browser-dynamic": "18.2.2",
+    "@angular/router": "18.2.2",
+    "@angular/service-worker": "18.2.2",
     "lodash": "^4.17.20",
-    "material-design-icons-iconfont": "^6.1.0",
-    "rxjs": "^7.5.0",
+    "rxjs": "7.8.1",
     "tailwindcss": "^3.4.4",
     "tslib": "^2.3.0",
     "zone.js": "0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.9",
-    "@angular-eslint/builder": "17.5.2",
-    "@angular-eslint/eslint-plugin": "17.5.2",
-    "@angular-eslint/eslint-plugin-template": "17.5.2",
-    "@angular-eslint/schematics": "17.5.2",
-    "@angular-eslint/template-parser": "17.5.2",
-    "@angular/cli": "17.3.9",
-    "@angular/compiler-cli": "17.3.12",
-    "@angular/language-service": "17.3.12",
-    "@biesbjerg/ngx-translate-extract": "^7.0.3",
-    "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
+    "@angular-devkit/build-angular": "18.2.2",
+    "@angular-eslint/builder": "18.3.0",
+    "@angular-eslint/eslint-plugin": "18.3.0",
+    "@angular-eslint/eslint-plugin-template": "18.3.0",
+    "@angular-eslint/schematics": "18.3.0",
+    "@angular-eslint/template-parser": "18.3.0",
+    "@angular/cli": "18.2.2",
+    "@angular/compiler-cli": "18.2.2",
+    "@angular/language-service": "18.2.2",
     "@compodoc/compodoc": "^1.1.11",
     "@ngneat/until-destroy": "^9.0.0",
-    "@ngx-rocket/scripts": "^5.2.2",
     "@types/jasmine": "^4.0.0",
     "@types/jasminewd2": "^2.0.8",
     "@types/lodash": "^4.14.184",
-    "@types/node": "^14.18.26",
-    "@typescript-eslint/eslint-plugin": "5.34.0",
-    "@typescript-eslint/parser": "5.34.0",
+    "@types/node": "18.19.47",
+    "@typescript-eslint/eslint-plugin": "8.3.0",
+    "@typescript-eslint/parser": "8.3.0",
     "daisyui": "^4.12.10",
     "eslint": "^8.3.0",
     "eslint-plugin-import": "latest",
     "eslint-plugin-jsdoc": "latest",
     "eslint-plugin-prefer-arrow": "latest",
-    "hads": "^3.0.0",
     "https-proxy-agent": "^5.0.0",
     "husky": "^8.0.1",
     "jasmine-core": "4.2.0",
