@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { TranslateModule } from '@ngx-translate/core';
 
 import { HomeRoutingModule } from './home-routing.module';
 import { HomeComponent } from './home.component';
@@ -9,7 +8,7 @@ import { WorkSectionComponent } from './components/work-section/work-section.com
 import { SharedModule } from '@app/shared/shared.module';
 
 @NgModule({
-  imports: [CommonModule, TranslateModule, HomeRoutingModule, SharedModule],
+  imports: [CommonModule, HomeRoutingModule, SharedModule],
   declarations: [HomeComponent, HeroSectionComponent, WorkSectionComponent],
   exports: [HeroSectionComponent, WorkSectionComponent],
 })
